+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Get upgrade notes from Sprockets 3.x to 4.x at https://github.com/rails/sprocket
 ## Master
 
 - Fix for Ruby 2.7 keyword arguments warning in `base.rb`. [#660](https://github.com/rails/sprockets/pull/660)
+- Adding new directive `depend_on_directory` [#668](https://github.com/rails/sprockets/pull/668)
 
 ## 4.0.0